chore: centralise embed validation checking
