+++ packages/components/src/utils/validation.ts
@@ -18,6 +18,42 @@ export const LINK_HREF_PATTERN =
 export const REF_HREF_PATTERN =
   `(${ALLOWED_URL_REGEXES.external})|(${ALLOWED_URL_REGEXES.internal})|(${ALLOWED_URL_REGEXES.files})|(${ALLOWED_URL_REGEXES.legacy})` as const
 
+// Validation for map-related embed URLs
+const isValidGoogleMapsEmbedUrl = (urlObject: URL) => {
+  return (
+    urlObject.hostname === "www.google.com" &&
+    urlObject.pathname === "/maps/embed"
+  )
+}
+
+const isValidOneMapEmbedUrl = (urlObject: URL) => {
+  return (
+    urlObject.hostname === "www.onemap.gov.sg" &&
+    urlObject.pathname === "/minimap/minimap.html"
+  )
+}
+
+export const isValidMapEmbedUrl = (url: string) => {
+  if (!url) {
+    return false
+  }
+
+  try {
+    const urlObject = new URL(url)
+
+    return (
+      (isValidGoogleMapsEmbedUrl(urlObject) ||
+        isValidOneMapEmbedUrl(urlObject)) &&
+      new RegExp(MAPS_EMBED_URL_PATTERN).test(url)
+    )
+  } catch (_) {
+    return false
+  }
+}
+
+// NOTE: This validation is still needed as this is the only validation method
+// that is supported inside the JSON schema. Components rely on the URL object
+// validation for better security.
 export const MAPS_EMBED_URL_REGEXES = {
   googlemaps: "^https://www\\.google\\.com/maps/embed?.*$",
   onemap: "^https://www\\.onemap\\.gov\\.sg/minimap/minimap\\.html.*$",
@@ -27,6 +63,39 @@ export const MAPS_EMBED_URL_PATTERN = Object.values(MAPS_EMBED_URL_REGEXES)
   .map((re) => `(${re})`)
   .join("|")
 
+// Validation for video-related embed URLs
+export const VALID_VIDEO_DOMAINS = {
+  youtube: [
+    "www.youtube.com",
+    "www.youtube-nocookie.com",
+    "youtube.com",
+    "youtube-nocookie.com",
+  ],
+  vimeo: ["player.vimeo.com"],
+  fbvideo: ["www.facebook.com"],
+}
+
+export const isValidVideoUrl = (url: string) => {
+  if (!url) {
+    return false
+  }
+
+  try {
+    const urlObject = new URL(url)
+    const allValidVideoDomains = Object.values(VALID_VIDEO_DOMAINS).flat()
+
+    return (
+      allValidVideoDomains.includes(urlObject.hostname) &&
+      new RegExp(VIDEO_EMBED_URL_PATTERN).test(url)
+    )
+  } catch (_) {
+    return false
+  }
+}
+
+// NOTE: This validation is still needed as this is the only validation method
+// that is supported inside the JSON schema. Components rely on the URL object
+// validation for better security.
 export const VIDEO_EMBED_URL_REGEXES = {
   fbvideo: "^https://www\\.facebook\\.com/plugins/video.php?.*$",
   vimeo: "^https://player\\.vimeo\\.com/video/.*$",
